requirements.txt
Adafruit-Blinka==8.45.0
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.71.0
Adafruit-PureIO==1.1.11
async-timeout==4.0.3
binho-host-adapter==0.1.6
colorzero==2.0
future==1.0.0
gpiozero==2.0.1
lxml==5.2.2
numpy==2.0.0
pyftdi==0.55.4
pymavlink==2.4.41
pyserial==3.5
python-decouple==3.8
pyusb==1.2.1
redis==5.0.7
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
